Fatten up and unminimize exeuntu and create exedev user with passwordless sudo
Containers are layered, and we want folks to have a nice experience,
so let's install way more common things. I experimented with installing
"everything", and that's fine too, but installing the built-in meta
packages is a good start.

We also create an exedev and give it sudo.

All commands have working man pages.

I haven't yet enabled using exedev as the default because there are a
few places that need to be updated to run sshd as root.

Co-Authored-By: sketch <hello@sketch.dev>
Change-ID: s6c404594bdc358f1k